split out ElementalStaticStiffnessMatrix and NodalStaticStiffnessMatrixComponents
