+++ src/gebt_poc/solver.cpp
@@ -13,317 +13,9 @@
 #include "src/gebt_poc/NodalElasticForces.hpp"
 #include "src/gebt_poc/SectionalMassMatrix.hpp"
 #include "src/gebt_poc/NodalInertialForces.hpp"
+#include "src/gebt_poc/NodalStaticStiffnessMatrixComponents.hpp"
 
 namespace openturbine::gebt_poc {
-
-void CalculateOMatrix(
-    View2D::const_type N_tilde, View2D::const_type M_tilde, View2D::const_type C11,
-    View2D::const_type C21, View2D::const_type values, View2D O_matrix
-) {
-    // non_zero_terms_part_1 = -N_tilde + [C11] * (x_0_prime_tilde + u_prime_tilde)
-    auto non_zero_terms_part_1 = View2D_3x3("non_zero_terms_part_1");
-    KokkosBlas::gemm("N", "N", 1., C11, values, 0., non_zero_terms_part_1);
-    KokkosBlas::axpy(-1., N_tilde, non_zero_terms_part_1);
-
-    // non_zero_terms_part_2 = -M_tilde + [C21] * (x_0_prime_tilde  + u_prime_tilde)
-    auto non_zero_terms_part_2 = View2D_3x3("non_zero_terms_part_2");
-    KokkosBlas::gemm("N", "N", 1., C21, values, 0., non_zero_terms_part_2);
-    KokkosBlas::axpy(-1., M_tilde, non_zero_terms_part_2);
-
-    // Assemble the O matrix
-    // [O]_6x6 = [
-    //     [0]_3x3      -N_tilde + [C11] * (x_0_prime_tilde + u_prime_tilde)
-    //     [0]_3x3      -M_tilde + [C21] * (x_0_prime_tilde  + u_prime_tilde)
-    // ]
-    Kokkos::deep_copy(O_matrix, 0.);
-    auto O_matrix_1 = Kokkos::subview(O_matrix, Kokkos::make_pair(0, 3), Kokkos::make_pair(3, 6));
-    Kokkos::deep_copy(O_matrix_1, non_zero_terms_part_1);
-    auto O_matrix_2 = Kokkos::subview(O_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(3, 6));
-    Kokkos::deep_copy(O_matrix_2, non_zero_terms_part_2);
-}
-
-void CalculatePMatrix(
-    View2D::const_type N_tilde, View2D::const_type C11, View2D::const_type C12,
-    View2D::const_type values, View2D P_matrix
-) {
-    // non_zero_terms_part_3 = (x_0_prime_tilde + u_prime_tilde)^T * [C11]
-    auto non_zero_terms_part_3 = View2D_3x3("non_zero_terms_part_3");
-    KokkosBlas::gemm("T", "N", 1., values, C11, 0., non_zero_terms_part_3);
-    KokkosBlas::axpy(1., N_tilde, non_zero_terms_part_3);
-
-    // non_zero_terms_part_4 = (x_0_prime_tilde  + u_prime_tilde)^T * [C12]
-    auto non_zero_terms_part_4 = View2D_3x3("non_zero_terms_part_4");
-    KokkosBlas::gemm("T", "N", 1., values, C12, 0., non_zero_terms_part_4);
-
-    // Assemble the P matrix
-    // [P]_6x6 = [
-    //                         [0]_3x3                                      [0]_3x3
-    //     N_tilde + (x_0_prime_tilde + u_prime_tilde)^T * [C11]     (x_0_prime_tilde  +
-    //     u_prime_tilde)^T * [C12]
-    // ]
-    Kokkos::deep_copy(P_matrix, 0.);
-    auto P_matrix_1 = Kokkos::subview(P_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(0, 3));
-    Kokkos::deep_copy(P_matrix_1, non_zero_terms_part_3);
-    auto P_matrix_2 = Kokkos::subview(P_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(3, 6));
-    Kokkos::deep_copy(P_matrix_2, non_zero_terms_part_4);
-}
-
-void CalculateQMatrix(
-    View2D::const_type N_tilde, View2D::const_type C11, View2D::const_type values, View2D Q_matrix
-) {
-    // Assemble the Q matrix
-    // [Q]_6x6 = [
-    //     [0]_3x3                                          [0]_3x3
-    //     [0]_3x3      (x_0_prime_tilde + u_prime_tilde)^T * (-N_tilde + [C11] * (x_0_prime_tilde  +
-    //     u_prime_tilde))
-    // ]
-    auto val = View2D_3x3("val");
-    KokkosBlas::gemm("N", "N", 1., C11, values, 0., val);
-    KokkosBlas::axpy(-1., N_tilde, val);
-
-    Kokkos::deep_copy(Q_matrix, 0.);
-    auto non_zero_term = Kokkos::subview(Q_matrix, Kokkos::make_pair(3, 6), Kokkos::make_pair(3, 6));
-    KokkosBlas::gemm("T", "N", 1., values, val, 0., non_zero_term);
-}
-
-void NodalStaticStiffnessMatrixComponents(
-    View1D_LieAlgebra::const_type elastic_force_fc,
-    View1D_LieGroup::const_type pos_vector_derivatives,
-    View1D_LieGroup::const_type gen_coords_derivatives, View2D_6x6::const_type sectional_stiffness,
-    View2D_6x6 O_matrix, View2D_6x6 P_matrix, View2D_6x6 Q_matrix
-) {
-    auto x0_prime_tilde = gen_alpha_solver::create_cross_product_matrix(
-        Kokkos::subview(pos_vector_derivatives, Kokkos::make_pair(0, 3))
-    );
-    auto u_prime_tilde = gen_alpha_solver::create_cross_product_matrix(
-        Kokkos::subview(gen_coords_derivatives, Kokkos::make_pair(0, 3))
-    );
-
-    auto values = View2D_3x3("values");
-    Kokkos::deep_copy(values, x0_prime_tilde);
-    KokkosBlas::axpy(1., u_prime_tilde, values);
-
-    auto N = Kokkos::subview(elastic_force_fc, Kokkos::make_pair(0, 3));
-    auto N_tilde = gen_alpha_solver::create_cross_product_matrix(N);
-    auto M = Kokkos::subview(elastic_force_fc, Kokkos::make_pair(3, 6));
-    auto M_tilde = gen_alpha_solver::create_cross_product_matrix(M);
-
-    auto C11 =
-        Kokkos::subview(sectional_stiffness, Kokkos::make_pair(0, 3), Kokkos::make_pair(0, 3));
-    auto C12 =
-        Kokkos::subview(sectional_stiffness, Kokkos::make_pair(0, 3), Kokkos::make_pair(3, 6));
-    auto C21 =
-        Kokkos::subview(sectional_stiffness, Kokkos::make_pair(3, 6), Kokkos::make_pair(0, 3));
-
-    // Calculate the O, P, and Q matrices
-    CalculateOMatrix(N_tilde, M_tilde, C11, C21, values, O_matrix);
-    CalculatePMatrix(N_tilde, C11, C12, values, P_matrix);
-    CalculateQMatrix(N_tilde, C11, values, Q_matrix);
-}
-
-void ElementalStaticStiffnessMatrix(
-    View1D::const_type position_vectors, View1D::const_type gen_coords,
-    const StiffnessMatrix& stiffness, const Quadrature& quadrature, View2D stiffness_matrix
-) {
-    const auto n_nodes = gen_coords.extent(0) / LieGroupComponents;
-    const auto order = n_nodes - 1;
-    const auto n_quad_pts = quadrature.GetNumberOfQuadraturePoints();
-
-    auto nodes = VectorFieldView("nodes", n_nodes);
-    for (std::size_t i = 0; i < n_nodes; ++i) {
-        auto index = i * LieGroupComponents;
-        Kokkos::deep_copy(
-            Kokkos::subview(nodes, i, Kokkos::ALL),
-            Kokkos::subview(position_vectors, Kokkos::make_pair(index, index + 3))
-        );
-    }
-
-    // Allocate Views for some required intermediate variables
-    auto gen_coords_qp = View1D_LieGroup("gen_coords_qp");
-    auto gen_coords_derivatives_qp = View1D_LieGroup("gen_coords_derivatives_qp");
-    auto position_vector_qp = View1D_LieGroup("position_vector_qp");
-    auto pos_vector_derivatives_qp = View1D_LieGroup("pos_vector_derivatives_qp");
-    auto curvature = View1D_Vector("curvature");
-    auto sectional_strain = View1D_LieAlgebra("sectional_strain");
-    auto sectional_stiffness = View2D_6x6("sectional_stiffness");
-    auto O_matrix = View2D_6x6("O_matrix");
-    auto P_matrix = View2D_6x6("P_matrix");
-    auto Q_matrix = View2D_6x6("Q_matrix");
-
-    Kokkos::deep_copy(stiffness_matrix, 0.);
-    for (size_t i = 0; i < n_nodes; ++i) {
-        for (size_t j = 0; j < n_nodes; ++j) {
-            for (size_t k = 0; k < n_quad_pts; ++k) {
-                // Calculate required interpolated values at the quadrature point
-                const auto q_pt = quadrature.GetQuadraturePoints()[k];
-                auto shape_function = LagrangePolynomial(order, q_pt);
-                auto shape_function_derivative = LagrangePolynomialDerivative(order, q_pt);
-                auto shape_function_vector = gen_alpha_solver::create_vector(shape_function);
-                auto shape_function_derivative_vector =
-                    gen_alpha_solver::create_vector(shape_function_derivative);
-
-                auto jacobian = CalculateJacobian(nodes, shape_function_derivative_vector);
-                InterpolateNodalValues(gen_coords, shape_function, gen_coords_qp);
-                InterpolateNodalValueDerivatives(
-                    gen_coords, shape_function_derivative, jacobian, gen_coords_derivatives_qp
-                );
-                InterpolateNodalValues(position_vectors, shape_function, position_vector_qp);
-                InterpolateNodalValueDerivatives(
-                    position_vectors, shape_function_derivative, jacobian, pos_vector_derivatives_qp
-                );
-
-                // Calculate the curvature and sectional strain
-                NodalCurvature(gen_coords_qp, gen_coords_derivatives_qp, curvature);
-                CalculateSectionalStrain(
-                    pos_vector_derivatives_qp, gen_coords_derivatives_qp, curvature, sectional_strain
-                );
-
-                // Calculate the sectional stiffness matrix in inertial basis
-                auto rotation_0 = gen_alpha_solver::EulerParameterToRotationMatrix(
-                    Kokkos::subview(position_vector_qp, Kokkos::make_pair(3, 7))
-                );
-                auto rotation = gen_alpha_solver::EulerParameterToRotationMatrix(
-                    Kokkos::subview(gen_coords_qp, Kokkos::make_pair(3, 7))
-                );
-                SectionalStiffness(stiffness, rotation_0, rotation, sectional_stiffness);
-
-                // Calculate elastic forces i.e. F^C and F^D vectors
-                auto elastic_forces_fc = View1D_LieAlgebra("elastic_forces_fc");
-                auto elastic_forces_fd = View1D_LieAlgebra("elastic_forces_fd");
-                NodalElasticForces(
-                    sectional_strain, rotation, pos_vector_derivatives_qp, gen_coords_derivatives_qp,
-                    sectional_stiffness, elastic_forces_fc, elastic_forces_fd
-                );
-
-                // Calculate the stiffness matrix components, i.e. O, P, and Q matrices
-                NodalStaticStiffnessMatrixComponents(
-                    elastic_forces_fc, pos_vector_derivatives_qp, gen_coords_derivatives_qp,
-                    sectional_stiffness, O_matrix, P_matrix, Q_matrix
-                );
-
-                const auto q_weight = quadrature.GetQuadratureWeights()[k];
-                Kokkos::parallel_for(
-                    Kokkos::MDRangePolicy<Kokkos::DefaultExecutionSpace, Kokkos::Rank<2>>(
-                        {0, 0}, {LieAlgebraComponents, LieAlgebraComponents}
-                    ),
-                    KOKKOS_LAMBDA(const size_t ii, const size_t jj) {
-                        stiffness_matrix(
-                            i * LieAlgebraComponents + ii, j * LieAlgebraComponents + jj
-                        ) += q_weight *
-                             (shape_function_vector(i) * P_matrix(ii, jj) *
-                                  shape_function_derivative_vector(j) +
-                              shape_function_vector(i) * Q_matrix(ii, jj) *
-                                  shape_function_vector(j) * jacobian +
-                              shape_function_derivative_vector(i) * sectional_stiffness(ii, jj) *
-                                  shape_function_derivative_vector(j) / jacobian +
-                              shape_function_derivative_vector(i) * O_matrix(ii, jj) *
-                                  shape_function_vector(j));
-                    }
-                );
-            }
-        }
-    }
-}
-
-void ElementalStaticStiffnessMatrix(
-    LieGroupFieldView::const_type position_vectors, LieGroupFieldView::const_type gen_coords,
-    const StiffnessMatrix& stiffness, const Quadrature& quadrature, View2D stiffness_matrix
-) {
-    const auto n_nodes = gen_coords.extent(0);
-    const auto order = n_nodes - 1;
-    const auto n_quad_pts = quadrature.GetNumberOfQuadraturePoints();
-
-    auto nodes = VectorFieldView("nodes", n_nodes);
-    Kokkos::deep_copy(
-        nodes, Kokkos::subview(position_vectors, Kokkos::ALL, Kokkos::make_pair(0, 3))
-    );
-
-    // Allocate Views for some required intermediate variables
-    auto gen_coords_qp = View1D_LieGroup("gen_coords_qp");
-    auto gen_coords_derivatives_qp = View1D_LieGroup("gen_coords_derivatives_qp");
-    auto position_vector_qp = View1D_LieGroup("position_vector_qp");
-    auto pos_vector_derivatives_qp = View1D_LieGroup("pos_vector_derivatives_qp");
-    auto curvature = View1D_Vector("curvature");
-    auto sectional_strain = View1D_LieAlgebra("sectional_strain");
-    auto sectional_stiffness = View2D_6x6("sectional_stiffness");
-    auto O_matrix = View2D_6x6("O_matrix");
-    auto P_matrix = View2D_6x6("P_matrix");
-    auto Q_matrix = View2D_6x6("Q_matrix");
-
-    Kokkos::deep_copy(stiffness_matrix, 0.);
-    for (size_t i = 0; i < n_nodes; ++i) {
-        for (size_t j = 0; j < n_nodes; ++j) {
-            for (size_t k = 0; k < n_quad_pts; ++k) {
-                // Calculate required interpolated values at the quadrature point
-                const auto q_pt = quadrature.GetQuadraturePoints()[k];
-                auto shape_function = LagrangePolynomial(order, q_pt);
-                auto shape_function_derivative = LagrangePolynomialDerivative(order, q_pt);
-                auto shape_function_vector = gen_alpha_solver::create_vector(shape_function);
-                auto shape_function_derivative_vector =
-                    gen_alpha_solver::create_vector(shape_function_derivative);
-
-                auto jacobian = CalculateJacobian(nodes, shape_function_derivative_vector);
-                InterpolateNodalValues(gen_coords, shape_function, gen_coords_qp);
-                InterpolateNodalValueDerivatives(
-                    gen_coords, shape_function_derivative, jacobian, gen_coords_derivatives_qp
-                );
-                InterpolateNodalValues(position_vectors, shape_function, position_vector_qp);
-                InterpolateNodalValueDerivatives(
-                    position_vectors, shape_function_derivative, jacobian, pos_vector_derivatives_qp
-                );
-
-                // Calculate the curvature and sectional strain
-                NodalCurvature(gen_coords_qp, gen_coords_derivatives_qp, curvature);
-                CalculateSectionalStrain(
-                    pos_vector_derivatives_qp, gen_coords_derivatives_qp, curvature, sectional_strain
-                );
-
-                // Calculate the sectional stiffness matrix in inertial basis
-                auto rotation_0 = gen_alpha_solver::EulerParameterToRotationMatrix(
-                    Kokkos::subview(position_vector_qp, Kokkos::make_pair(3, 7))
-                );
-                auto rotation = gen_alpha_solver::EulerParameterToRotationMatrix(
-                    Kokkos::subview(gen_coords_qp, Kokkos::make_pair(3, 7))
-                );
-                SectionalStiffness(stiffness, rotation_0, rotation, sectional_stiffness);
-
-                // Calculate elastic forces i.e. F^C and F^D vectors
-                auto elastic_forces_fc = View1D_LieAlgebra("elastic_forces_fc");
-                auto elastic_forces_fd = View1D_LieAlgebra("elastic_forces_fd");
-                NodalElasticForces(
-                    sectional_strain, rotation, pos_vector_derivatives_qp, gen_coords_derivatives_qp,
-                    sectional_stiffness, elastic_forces_fc, elastic_forces_fd
-                );
-
-                // Calculate the stiffness matrix components, i.e. O, P, and Q matrices
-                NodalStaticStiffnessMatrixComponents(
-                    elastic_forces_fc, pos_vector_derivatives_qp, gen_coords_derivatives_qp,
-                    sectional_stiffness, O_matrix, P_matrix, Q_matrix
-                );
-
-                const auto q_weight = quadrature.GetQuadratureWeights()[k];
-                Kokkos::parallel_for(
-                    Kokkos::MDRangePolicy<Kokkos::DefaultExecutionSpace, Kokkos::Rank<2>>(
-                        {0, 0}, {LieAlgebraComponents, LieAlgebraComponents}
-                    ),
-                    KOKKOS_LAMBDA(const size_t ii, const size_t jj) {
-                        stiffness_matrix(
-                            i * LieAlgebraComponents + ii, j * LieAlgebraComponents + jj
-                        ) += q_weight *
-                             (shape_function_vector(i) * P_matrix(ii, jj) *
-                                  shape_function_derivative_vector(j) +
-                              shape_function_vector(i) * Q_matrix(ii, jj) *
-                                  shape_function_vector(j) * jacobian +
-                              shape_function_derivative_vector(i) * sectional_stiffness(ii, jj) *
-                                  shape_function_derivative_vector(j) / jacobian +
-                              shape_function_derivative_vector(i) * O_matrix(ii, jj) *
-                                  shape_function_vector(j));
-                    }
-                );
-            }
-        }
-    }
-}
 
 void NodalGyroscopicMatrix(
     View1D_LieAlgebra::const_type velocity, const MassMatrix& sectional_mass_matrix,
