+++ tests/unit_tests/gebt_poc/test_NodalStaticStiffnessMatrixComponents.cpp
@@ -1,0 +1,90 @@
+#include <gtest/gtest.h>
+
+#include "src/gebt_poc/NodalStaticStiffnessMatrixComponents.hpp"
+
+#include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
+
+namespace openturbine::gebt_poc {
+    
+TEST(SolverTest, NodalStaticStiffnessMatrixComponents) {
+    auto elastic_force_fc = gen_alpha_solver::create_vector(
+        {0.1023527958818833, 0.1512321779691288, 0.2788924951018168, 0.4003985306163255,
+         0.3249298550145402, 0.5876343707088096}
+    );
+
+    auto position_vector_derivatives = gen_alpha_solver::create_vector(
+        {0.924984344499876, -0.3417491071948322, 0.16616711516322974, 0.023197240723436388,
+         0.0199309451611758, 0.0569650074322926, 0.}
+    );
+
+    auto gen_coords_derivatives = gen_alpha_solver::create_vector(
+        {0.0009414876868372689, -0.0009055519814222231, 0.000948674827920281,
+         -0.000011768592509980857, 0.009249835939573452, 0., 0.}
+    );
+
+    auto stiffness = gen_alpha_solver::create_matrix(
+        {{1.3197900000789533, 1.9500660871457987, 3.596184383250417, 5.162946182374572,
+          4.189813963364337, 7.577262149821259},
+         {1.9500660871457987, 2.881335473074227, 5.313570498700886, 7.628551708533343,
+          6.190692550267803, 11.19584324089132},
+         {3.5961843832504163, 5.313570498700887, 9.798939314254936, 14.068076308736293,
+          11.416470455810389, 20.64664212439219},
+         {5.16294618237457, 7.628551708533343, 14.068076308736295, 20.197162639890838,
+          16.390322707186375, 29.641834448609774},
+         {4.189813963364336, 6.190692550267804, 11.416470455810385, 16.39032270718637,
+          13.301010802137164, 24.054825962838},
+         {7.577262149821259, 11.19584324089132, 20.64664212439219, 29.64183444860977,
+          24.054825962838, 43.50305858028866}}
+    );
+
+    auto O_matrix = Kokkos::View<double**>("O_matrix", 6, 6);
+    auto P_matrix = Kokkos::View<double**>("P_matrix", 6, 6);
+    auto Q_matrix = Kokkos::View<double**>("Q_matrix", 6, 6);
+
+    NodalStaticStiffnessMatrixComponents(
+        elastic_force_fc, position_vector_derivatives, gen_coords_derivatives, stiffness, O_matrix,
+        P_matrix, Q_matrix
+    );
+
+    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
+        O_matrix,
+        {
+            {0., 0., 0., 1.5581361688659614, 3.3881347643742066, -2.409080935189973},  // row 1
+            {0., 0., 0., 2.023343846951859, 4.594085351204066, -3.233749380295583},    // row 2
+            {0., 0., 0., 4.396865920548022, 8.369759049055988, -6.152221520070027},    // row 3
+            {0., 0., 0., 6.095343338095462, 12.750819804192329, -9.15751050858455},    // row 4
+            {0., 0., 0., 4.358834898453007, 9.870620837027674, -6.767382896430996},    // row 5
+            {0., 0., 0., 9.270600163100875, 17.450580610135344, -12.962904649290419}   // row 6
+        }
+    );
+
+    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
+        P_matrix,
+        {
+            {0., 0., 0., 0., 0., 0.},  // row 1
+            {0., 0., 0., 0., 0., 0.},  // row 2
+            {0., 0., 0., 0., 0., 0.},  // row 3
+            {1.558136168865961, 2.023343846951859, 4.396865920548022, 6.095343338095461,
+             4.946469269161818, 8.945670308086335},  // row 4
+            {3.388134764374206, 4.594085351204066, 8.369759049055988, 12.163185433483518,
+             9.870620837027678, 17.85097914075167},  // row 5
+            {-2.409080935189973, -3.233749380295583, -6.152221520070027, -8.83258065357001,
+             -7.16778142704732, -12.962904649290419}  // row 6
+        }
+    );
+
+    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
+        Q_matrix,
+        {
+            {0., 0., 0., 0., 0., 0.},                                                  // row 1
+            {0., 0., 0., 0., 0., 0.},                                                  // row 2
+            {0., 0., 0., 0., 0., 0.},                                                  // row 3
+            {0., 0., 0., 1.844739298856163, 3.6356811370948883, -2.648497950457901},   // row 4
+            {0., 0., 0., 3.8107825797230066, 7.1835652949545645, -5.293905367130955},  // row 5
+            {0., 0., 0., -2.4073689531817264, -5.414742464880276, 3.8196948928089887}  // row 6
+
+        }
+    );
+}
+
+}
