+++ tests/unit_tests/gebt_poc/test_solver.cpp
@@ -590,301 +590,6 @@ TEST(SolverTest, ElementalInertialMatrices2D) {
     );
 }
 
-TEST(SolverTest, NodalStaticStiffnessMatrixComponents) {
-    auto elastic_force_fc = gen_alpha_solver::create_vector(
-        {0.1023527958818833, 0.1512321779691288, 0.2788924951018168, 0.4003985306163255,
-         0.3249298550145402, 0.5876343707088096}
-    );
-
-    auto position_vector_derivatives = gen_alpha_solver::create_vector(
-        {0.924984344499876, -0.3417491071948322, 0.16616711516322974, 0.023197240723436388,
-         0.0199309451611758, 0.0569650074322926, 0.}
-    );
-
-    auto gen_coords_derivatives = gen_alpha_solver::create_vector(
-        {0.0009414876868372689, -0.0009055519814222231, 0.000948674827920281,
-         -0.000011768592509980857, 0.009249835939573452, 0., 0.}
-    );
-
-    auto stiffness = gen_alpha_solver::create_matrix(
-        {{1.3197900000789533, 1.9500660871457987, 3.596184383250417, 5.162946182374572,
-          4.189813963364337, 7.577262149821259},
-         {1.9500660871457987, 2.881335473074227, 5.313570498700886, 7.628551708533343,
-          6.190692550267803, 11.19584324089132},
-         {3.5961843832504163, 5.313570498700887, 9.798939314254936, 14.068076308736293,
-          11.416470455810389, 20.64664212439219},
-         {5.16294618237457, 7.628551708533343, 14.068076308736295, 20.197162639890838,
-          16.390322707186375, 29.641834448609774},
-         {4.189813963364336, 6.190692550267804, 11.416470455810385, 16.39032270718637,
-          13.301010802137164, 24.054825962838},
-         {7.577262149821259, 11.19584324089132, 20.64664212439219, 29.64183444860977,
-          24.054825962838, 43.50305858028866}}
-    );
-
-    auto O_matrix = Kokkos::View<double**>("O_matrix", 6, 6);
-    auto P_matrix = Kokkos::View<double**>("P_matrix", 6, 6);
-    auto Q_matrix = Kokkos::View<double**>("Q_matrix", 6, 6);
-
-    NodalStaticStiffnessMatrixComponents(
-        elastic_force_fc, position_vector_derivatives, gen_coords_derivatives, stiffness, O_matrix,
-        P_matrix, Q_matrix
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
-        O_matrix,
-        {
-            {0., 0., 0., 1.5581361688659614, 3.3881347643742066, -2.409080935189973},  // row 1
-            {0., 0., 0., 2.023343846951859, 4.594085351204066, -3.233749380295583},    // row 2
-            {0., 0., 0., 4.396865920548022, 8.369759049055988, -6.152221520070027},    // row 3
-            {0., 0., 0., 6.095343338095462, 12.750819804192329, -9.15751050858455},    // row 4
-            {0., 0., 0., 4.358834898453007, 9.870620837027674, -6.767382896430996},    // row 5
-            {0., 0., 0., 9.270600163100875, 17.450580610135344, -12.962904649290419}   // row 6
-        }
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
-        P_matrix,
-        {
-            {0., 0., 0., 0., 0., 0.},  // row 1
-            {0., 0., 0., 0., 0., 0.},  // row 2
-            {0., 0., 0., 0., 0., 0.},  // row 3
-            {1.558136168865961, 2.023343846951859, 4.396865920548022, 6.095343338095461,
-             4.946469269161818, 8.945670308086335},  // row 4
-            {3.388134764374206, 4.594085351204066, 8.369759049055988, 12.163185433483518,
-             9.870620837027678, 17.85097914075167},  // row 5
-            {-2.409080935189973, -3.233749380295583, -6.152221520070027, -8.83258065357001,
-             -7.16778142704732, -12.962904649290419}  // row 6
-        }
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
-        Q_matrix,
-        {
-            {0., 0., 0., 0., 0., 0.},                                                  // row 1
-            {0., 0., 0., 0., 0., 0.},                                                  // row 2
-            {0., 0., 0., 0., 0., 0.},                                                  // row 3
-            {0., 0., 0., 1.844739298856163, 3.6356811370948883, -2.648497950457901},   // row 4
-            {0., 0., 0., 3.8107825797230066, 7.1835652949545645, -5.293905367130955},  // row 5
-            {0., 0., 0., -2.4073689531817264, -5.414742464880276, 3.8196948928089887}  // row 6
-
-        }
-    );
-}
-
-TEST(SolverTest, ElementalStaticStiffnessMatrixWithZeroValues) {
-    // 5 nodes on a line
-    auto position_vectors = gen_alpha_solver::create_vector({
-        0., 0., 0., 1., 0., 0., 0.,  // node 1
-        1., 0., 0., 1., 0., 0., 0.,  // node 2
-        2., 0., 0., 1., 0., 0., 0.,  // node 3
-        3., 0., 0., 1., 0., 0., 0.,  // node 4
-        4., 0., 0., 1., 0., 0., 0.   // node 5
-    });
-
-    // zero displacement and rotation
-    auto generalized_coords = gen_alpha_solver::create_vector({
-        0., 0., 0., 0., 0., 0., 0.,  // node 1
-        0., 0., 0., 0., 0., 0., 0.,  // node 2
-        0., 0., 0., 0., 0., 0., 0.,  // node 3
-        0., 0., 0., 0., 0., 0., 0.,  // node 4
-        0., 0., 0., 0., 0., 0., 0.   // node 5
-    });
-
-    // identity stiffness matrix
-    auto stiffness = gen_alpha_solver::create_matrix({
-        {0., 0., 0., 0., 0., 0.},  // row 1
-        {0., 0., 0., 0., 0., 0.},  // row 2
-        {0., 0., 0., 0., 0., 0.},  // row 3
-        {0., 0., 0., 0., 0., 0.},  // row 4
-        {0., 0., 0., 0., 0., 0.},  // row 5
-        {0., 0., 0., 0., 0., 0.}   // row 6
-    });
-
-    // 7-point Gauss-Legendre quadrature
-    auto quadrature = UserDefinedQuadrature(
-        {-0.9491079123427585, -0.7415311855993945, -0.4058451513773972, 0., 0.4058451513773972,
-         0.7415311855993945, 0.9491079123427585},
-        {0.1294849661688697, 0.2797053914892766, 0.3818300505051189, 0.4179591836734694,
-         0.3818300505051189, 0.2797053914892766, 0.1294849661688697}
-    );
-
-    auto iteration_matrix = Kokkos::View<double[30][30]>("iteration_matrix");
-    ElementalStaticStiffnessMatrix(
-        position_vectors, generalized_coords, StiffnessMatrix(stiffness), quadrature,
-        iteration_matrix
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(iteration_matrix, zeros_30x30);
-}
-
-TEST(SolverTest, ElementalStaticStiffnessMatrixWithNonZeroValues) {
-    auto position_vectors = gen_alpha_solver::create_vector(
-        {// node 1
-         0., 0., 0., 0.9778215200524469, -0.01733607539094763, -0.09001900002195001,
-         -0.18831121859148398,
-         // node 2
-         0.8633658232300573, -0.25589826392541715, 0.1130411210682743, 0.9950113028068008,
-         -0.002883848832932071, -0.030192109815745303, -0.09504013471947484,
-         // node 3
-         2.5, -0.25, 0., 0.9904718430204884, -0.009526411091536478, 0.09620741150793366,
-         0.09807604012323785,
-         // node 4
-         4.136634176769943, 0.39875540678255983, -0.5416125496397027, 0.9472312341234699,
-         -0.049692141629315074, 0.18127630174800594, 0.25965858850765167,
-         // node 5
-         5., 1., -1., 0.9210746582719719, -0.07193653093139739, 0.20507529985516368,
-         0.32309554437664584}
-    );
-
-    auto generalized_coords = gen_alpha_solver::create_vector(
-        {// node 1
-         0., 0., 0., 1., 0., 0., 0.,
-         // node 2
-         0.0029816021788868583, -0.0024667594949430213, 0.0030845707156756256, 0.9999627302042724,
-         0.008633550973807838, 0., 0.,
-         // node 3
-         0.025, -0.0125, 0.027500000000000004, 0.9996875162757026, 0.024997395914712332, 0., 0.,
-         // node 4
-         0.06844696924968456, -0.011818954790771264, 0.07977257214146723, 0.9991445348823056,
-         0.04135454527402519, 0., 0.,
-         // node 5
-         0.1, 0., 0.12, 0.9987502603949662, 0.049979169270678324, 0., 0.}
-    );
-
-    auto stiffness = StiffnessMatrix(gen_alpha_solver::create_matrix({
-        {1., 2., 3., 4., 5., 6.},       // row 1
-        {2., 4., 6., 8., 10., 12.},     // row 2
-        {3., 6., 9., 12., 15., 18.},    // row 3
-        {4., 8., 12., 16., 20., 24.},   // row 4
-        {5., 10., 15., 20., 25., 30.},  // row 5
-        {6., 12., 18., 24., 30., 36.}   // row 6
-    }));
-
-    auto quadrature_points =
-        std::vector<double>{-0.9491079123427585, -0.7415311855993945, -0.4058451513773972, 0.,
-                            0.4058451513773972,  0.7415311855993945,  0.9491079123427585};
-    auto quadrature_weights = std::vector<double>{
-        0.1294849661688697, 0.2797053914892766, 0.3818300505051189, 0.4179591836734694,
-        0.3818300505051189, 0.2797053914892766, 0.1294849661688697};
-    auto quadrature = UserDefinedQuadrature(quadrature_points, quadrature_weights);
-
-    auto iteration_matrix = Kokkos::View<double[30][30]>("iteration_matrix");
-    ElementalStaticStiffnessMatrix(
-        position_vectors, generalized_coords, stiffness, quadrature, iteration_matrix
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
-        iteration_matrix, expected_iteration_matrix
-    );
-}
-
-TEST(SolverTest, ElementalStaticStiffnessMatrixWithZeroValues2D) {
-    // 5 nodes on a line
-    auto position_vectors = gen_alpha_solver::create_matrix({
-        {0., 0., 0., 1., 0., 0., 0.},  // node 1
-        {1., 0., 0., 1., 0., 0., 0.},  // node 2
-        {2., 0., 0., 1., 0., 0., 0.},  // node 3
-        {3., 0., 0., 1., 0., 0., 0.},  // node 4
-        {4., 0., 0., 1., 0., 0., 0.}   // node 5
-    });
-
-    // zero displacement and rotation
-    auto generalized_coords = gen_alpha_solver::create_matrix({
-        {0., 0., 0., 0., 0., 0., 0.},  // node 1
-        {0., 0., 0., 0., 0., 0., 0.},  // node 2
-        {0., 0., 0., 0., 0., 0., 0.},  // node 3
-        {0., 0., 0., 0., 0., 0., 0.},  // node 4
-        {0., 0., 0., 0., 0., 0., 0.}   // node 5
-    });
-
-    // identity stiffness matrix
-    auto stiffness = gen_alpha_solver::create_matrix({
-        {0., 0., 0., 0., 0., 0.},  // row 1
-        {0., 0., 0., 0., 0., 0.},  // row 2
-        {0., 0., 0., 0., 0., 0.},  // row 3
-        {0., 0., 0., 0., 0., 0.},  // row 4
-        {0., 0., 0., 0., 0., 0.},  // row 5
-        {0., 0., 0., 0., 0., 0.}   // row 6
-    });
-
-    // 7-point Gauss-Legendre quadrature
-    auto quadrature = UserDefinedQuadrature(
-        {-0.9491079123427585, -0.7415311855993945, -0.4058451513773972, 0., 0.4058451513773972,
-         0.7415311855993945, 0.9491079123427585},
-        {0.1294849661688697, 0.2797053914892766, 0.3818300505051189, 0.4179591836734694,
-         0.3818300505051189, 0.2797053914892766, 0.1294849661688697}
-    );
-
-    auto iteration_matrix = Kokkos::View<double[30][30]>("iteration_matrix");
-    ElementalStaticStiffnessMatrix(
-        position_vectors, generalized_coords, StiffnessMatrix(stiffness), quadrature,
-        iteration_matrix
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(iteration_matrix, zeros_30x30);
-}
-
-TEST(SolverTest, ElementalStaticStiffnessMatrixWithNonZeroValues2D) {
-    auto position_vectors = gen_alpha_solver::create_matrix(
-        {// node 1
-         {0., 0., 0., 0.9778215200524469, -0.01733607539094763, -0.09001900002195001,
-          -0.18831121859148398},
-         // node 2
-         {0.8633658232300573, -0.25589826392541715, 0.1130411210682743, 0.9950113028068008,
-          -0.002883848832932071, -0.030192109815745303, -0.09504013471947484},
-         // node 3
-         {2.5, -0.25, 0., 0.9904718430204884, -0.009526411091536478, 0.09620741150793366,
-          0.09807604012323785},
-         // node 4
-         {4.136634176769943, 0.39875540678255983, -0.5416125496397027, 0.9472312341234699,
-          -0.049692141629315074, 0.18127630174800594, 0.25965858850765167},
-         // node 5
-         {5., 1., -1., 0.9210746582719719, -0.07193653093139739, 0.20507529985516368,
-          0.32309554437664584}}
-    );
-
-    auto generalized_coords = gen_alpha_solver::create_matrix(
-        {// node 1
-         {0., 0., 0., 1., 0., 0., 0.},
-         // node 2
-         {0.0029816021788868583, -0.0024667594949430213, 0.0030845707156756256, 0.9999627302042724,
-          0.008633550973807838, 0., 0.},
-         // node 3
-         {0.025, -0.0125, 0.027500000000000004, 0.9996875162757026, 0.024997395914712332, 0., 0.},
-         // node 4
-         {0.06844696924968456, -0.011818954790771264, 0.07977257214146723, 0.9991445348823056,
-          0.04135454527402519, 0., 0.},
-         // node 5
-         {0.1, 0., 0.12, 0.9987502603949662, 0.049979169270678324, 0., 0.}}
-    );
-
-    auto stiffness = StiffnessMatrix(gen_alpha_solver::create_matrix({
-        {1., 2., 3., 4., 5., 6.},       // row 1
-        {2., 4., 6., 8., 10., 12.},     // row 2
-        {3., 6., 9., 12., 15., 18.},    // row 3
-        {4., 8., 12., 16., 20., 24.},   // row 4
-        {5., 10., 15., 20., 25., 30.},  // row 5
-        {6., 12., 18., 24., 30., 36.}   // row 6
-    }));
-
-    auto quadrature_points =
-        std::vector<double>{-0.9491079123427585, -0.7415311855993945, -0.4058451513773972, 0.,
-                            0.4058451513773972,  0.7415311855993945,  0.9491079123427585};
-    auto quadrature_weights = std::vector<double>{
-        0.1294849661688697, 0.2797053914892766, 0.3818300505051189, 0.4179591836734694,
-        0.3818300505051189, 0.2797053914892766, 0.1294849661688697};
-    auto quadrature = UserDefinedQuadrature(quadrature_points, quadrature_weights);
-
-    auto iteration_matrix = Kokkos::View<double[30][30]>("iteration_matrix");
-    ElementalStaticStiffnessMatrix(
-        position_vectors, generalized_coords, stiffness, quadrature, iteration_matrix
-    );
-
-    openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
-        iteration_matrix, expected_iteration_matrix
-    );
-}
-
 TEST(SolverTest, ElementalConstraintForcesResidual) {
     auto generalized_coords = gen_alpha_solver::create_vector(
         {0.1, 0., 0.12, 0.9987502603949662, 0.049979169270678324, 0., 0.}
